requirements.txt
openai
pandas
langchain
schedule
python-dotenv